test just an extraction without modifications
